+++ src/css/Product.css
@@ -1,0 +1,105 @@
+.content-body {
+  min-height: 330px;
+  min-width: 320px;
+  margin-left: 318px;
+  margin-right: 48px;
+  transition: margin-left 0.3s;
+}
+.content-container {
+  overflow: visible;
+}
+.content-container .nav {
+  width: 1537px;
+  top: 0px;
+  position: static;
+  padding: 10px 0;
+  background: #fff;
+  margin: 0 auto;
+}
+.link-area {
+  position: relative;
+  width: 100%;
+  text-align: center;
+  background-color: white;
+}
+.first-menu {
+  height: 100%;
+  margin: 0 auto;
+}
+.first-menu li.active {
+  background: #111;
+  border: 1px solid #111;
+}
+.first-menu li {
+  display: inline-block;
+  height: 100%;
+  margin-right: 10px;
+  padding: 8px 16px;
+  background: #e9e9e9;
+  border-radius: 5px;
+  border: 1px solid #e9e9e9;
+}
+.first-menu li.active a {
+  color: #fff;
+}
+.first-menu li a {
+  color: #787878;
+}
+
+.nr-container {
+  padding-left: 0;
+  padding-right: 0;
+}
+.item-list-wrap {
+  display: flex;
+  flex-wrap: wrap;
+}
+.item-list-wrap li {
+  width: 32%;
+  display: inline-block;
+}
+.item-list {
+  padding-left: 8px;
+  padding-right: 8px;
+}
+.product-item {
+  overflow: hidden;
+}
+.product-img {
+  position: relative;
+  z-index: 2;
+  background: whitesmoke;
+  width: 100%;
+}
+.product-img img {
+  width: 100%;
+  height: 100%;
+  vertical-align: top;
+}
+.product-info {
+  text-align: left;
+  padding-top: 12px;
+}
+.product-info .product-display {
+  display: flex;
+  justify-content: space-between;
+  word-break: keep-all;
+  line-height: 1.6em;
+  letter-spacing: -0.5px;
+}
+.product-info-area {
+  width: 66.6%;
+  padding-right: 15px;
+}
+.product-price {
+  padding-right: 10px;
+}
+.product-display-name {
+  line-height: 1.4em;
+}
+.item-color-opt {
+  padding-top: 3px;
+}
+.product-display-price {
+  white-space: nowrap;
+}
