+++ src/pages/NR_page.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import Sidebar from "../components/Sidebar";
-import Main from "../components/Main";
+import Product from "../components/Product";
 
 export default function NR_page() {
-    return (
-        <>
-            <Sidebar/>
-            <Main/>
-        </>
-    )
+  return (
+    <>
+      <Sidebar />
+      <Product />
+    </>
+  );
 }
